requirements.txt
aiofiles==22.1.0
black==22.10.0
blinker==1.5
click==8.1.3
colorama==0.4.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hypercorn==0.14.3
hyperframe==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.31
pathspec==0.10.2
platformdirs==2.6.0
priority==2.0.0
protobuf==3.20.1
quart==0.18.3
toml==0.10.2
tomli==2.0.1
Werkzeug==2.2.2
wsproto==1.2.0
